Centralize definition of test main and include cstdint in several headers
Removed CATCH_CONFIG_MAIN from individual test files and consolidated it into a single test_main.cc.
This simplifies the test configuration and reduces the number of files defining the main function for the Catch2 test suite.
Additionally, the cstdint header was included in several algorithm and data structure headers to ensure the int64_t type is recognized. A new Bubble sort test was also added.

+++ src/algorithms/dynamic_programming/fib.h
@@ -4,6 +4,8 @@
 #ifdef __cplusplus
 #include <iostream>
 #include <vector>
+#include <math.h>
+#include <cstdint>
 #endif
 
 /**
@@ -62,6 +64,6 @@ int64_t fibonacci_bottom_up(int64_t n) {
  */
 int64_t fibonacci_binet(int64_t n) {
   double phi = (std::sqrt(5) + 1) / 2;
-  return std::round(std::pow(phi, n) / sqrt(5));
+  return (int64_t) std::round(std::pow(phi, n) / sqrt(5));
 }
 #endif
